--- src/main/java/com/shy_polarbear/server/domain/user/dto/TokenRequest.java
@@ -1,10 +1,0 @@
-package com.shy_polarbear.server.domain.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class TokenRequest {
-    private String accessToken;
-}
