+++ src/main/java/com/shy_polarbear/server/domain/user/service/AuthService.java
@@ -1,14 +1,14 @@
 package com.shy_polarbear.server.domain.user.service;
 
-import com.shy_polarbear.server.global.config.jwt.JwtDto;
-import com.shy_polarbear.server.global.config.jwt.RefreshToken;
-import com.shy_polarbear.server.global.config.jwt.RefreshTokenRepository;
-import com.shy_polarbear.server.global.config.security.PrincipalDetails;
-import com.shy_polarbear.server.domain.user.dto.*;
-import com.shy_polarbear.server.domain.user.dto.JoinRequest;
-import com.shy_polarbear.server.domain.user.dto.SocialLoginRequest;
+import com.shy_polarbear.server.domain.user.dto.auth.LogoutResponse;
+import com.shy_polarbear.server.global.auth.jwt.JwtDto;
+import com.shy_polarbear.server.global.auth.jwt.RefreshToken;
+import com.shy_polarbear.server.global.auth.jwt.RefreshTokenRepository;
+import com.shy_polarbear.server.global.auth.security.PrincipalDetails;
+import com.shy_polarbear.server.domain.user.dto.auth.JoinRequest;
+import com.shy_polarbear.server.domain.user.dto.auth.SocialLoginRequest;
 import com.shy_polarbear.server.domain.user.exception.AuthException;
-import com.shy_polarbear.server.global.config.jwt.JwtProvider;
+import com.shy_polarbear.server.global.auth.jwt.JwtProvider;
 import com.shy_polarbear.server.domain.user.exception.UserException;
 import com.shy_polarbear.server.domain.user.model.User;
 import com.shy_polarbear.server.domain.user.model.UserRole;
@@ -36,6 +36,7 @@ public class AuthService {
     private final KakaoProvider kakaoProvider;
     private final PasswordEncoder passwordEncoder;
     private final AuthenticationManager authenticationManager;
+    private final UserService userService;
 
     //DB에 유저가 있다면 로그인 처리, 없다면 회원가입 유도
     public JwtDto authLogin(SocialLoginRequest socialLoginRequest) {
@@ -57,16 +58,16 @@ public class AuthService {
         String providerId = userInfo.getId();
 
         //이미 가입된 유저인지 확인
-        checkDuplicationUser(providerId);
+        userService.checkDuplicationUser(providerId);
 
         //닉네임 중복 검증
-        checkDuplicationNickName(joinRequest.getNickName());
+        userService.checkDuplicationNickName(joinRequest.getNickName());
 
         //유저 저장
         User joinUser = User.createUser(joinRequest.getNickName(), joinRequest.getEmail(),
                 joinRequest.getProfileImage(), joinRequest.getPhoneNumber(),
                 UserRole.ROLE_USR, providerId, ProviderType.KAKAO.getValue(), passwordEncoder);
-        userRepository.save(joinUser);
+        userService.save(joinUser);
 
         //로그인 (유저 인증)
         JwtDto issuedToken = authorizeUser(providerId);
@@ -79,18 +80,6 @@ public class AuthService {
         SecurityContextHolder.getContext().setAuthentication(authenticated);
         PrincipalDetails principal = (PrincipalDetails) authenticated.getPrincipal();
         return jwtProvider.issue(principal.getUser());
-    }
-
-    public void checkDuplicationNickName(String nickName) {
-        if (userRepository.existsByNickName(nickName)) {
-            throw new UserException(ExceptionStatus.NICKNAME_DUPLICATION);
-        }
-    }
-
-    public void checkDuplicationUser(String providerId) {
-        if (userRepository.existsByProviderId(providerId)) {
-            throw new UserException(ExceptionStatus.USER_ALREADY_EXISTS);
-        }
     }
 
     // refresh token 삭제하는 방식 사용
