+++ src/main/java/com/shy_polarbear/server/global/auth/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.global.config.jwt;
+package com.shy_polarbear.server.global.auth.jwt;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.shy_polarbear.server.global.common.dto.ApiResponse;
